+++ resources/views/layouts/compact.blade.php
@@ -1,0 +1,6 @@
+<img src="{{ $image }}"
+     alt="sample"
+     class="mw-100 d-block img-fluid">
+<span class="small text-muted mt-1 mb-0">
+   {{ $id }}
+</span>